requirements.txt
meshtastic>=2.2.0
pypubsub>=4.0.3
